requirements.txt
crewai>=0.95.0
python-dotenv>=1.0.1
